refactor: move callbacks to state_machine.c
