+++ gripper_17.X/src/main.c
@@ -1,5 +1,6 @@
 #include "system_init.h"
 #include "gripper.h"
+#include "state_machine.h"
 
 
 uint8_t Can0MessageRAM[CAN0_MESSAGE_RAM_CONFIG_SIZE]
@@ -9,20 +10,16 @@ static uint16_t adc_result_array[TRANSFER_SIZE];
 static volatile uint32_t events;
 
 
-
-static void callback_init(void);
-void can_rx_callback(uintptr_t context);
-void dmac_channel0_callback(DMAC_TRANSFER_EVENT returned_evnt,
-                            uintptr_t MyDmacContext);
-void tc0_callback(TC_TIMER_STATUS status, uintptr_t context);
-void tc1_callback(TC_TIMER_STATUS status, uintptr_t context);
-
 int main(void) {
     system_init();
 
     CAN0_MessageRAMConfigSet(Can0MessageRAM);
 
-    callback_init();
+    CAN0_RxCallbackRegister(can_rx_callback, (uintptr_t) &rx_frame,
+                            CAN_MSG_ATTR_RX_FIFO0);
+    DMAC_ChannelCallbackRegister(DMAC_CHANNEL_0, dmac_channel0_callback, 0);
+    TC0_TimerCallbackRegister(tc0_callback, (uintptr_t)&events);
+    TC1_TimerCallbackRegister(tc1_callback, (uintptr_t)&events);
 
 
     DMAC_ChannelTransfer(DMAC_CHANNEL_0, (const void*)&ADC0_REGS->ADC_RESULT,
@@ -63,106 +60,3 @@ int main(void) {
     return EXIT_FAILURE;
 }
 
-
-void can_rx_callback(uintptr_t context) {
-    if (CAN0_ErrorGet()) {
-        return;
-    }
-    switch (rx_frame.id) {
-        case STOP_GRIPPER:
-            stop_gripper();
-            break;
-        case START_GRIPPER:
-            start_gripper();
-            break;
-        case SET_PWM:
-            set_servos_pwm(rx_frame.buf);
-            WDT_Clear();
-            break;
-        case RESET_MCU:
-            NVIC_SystemReset();
-            break;
-        default:
-            break;
-    }
-}
-
-void tc0_callback(TC_TIMER_STATUS status, uintptr_t context) {
-    events |= EVENT_READ_ENCODER;
-}
-
-void tc1_callback(TC_TIMER_STATUS status, uintptr_t context) {
-    events |= EVENT_TRANSMIT_ANGLES;
-}
-
-void dmac_channel0_callback(DMAC_TRANSFER_EVENT returned_evnt,
-                            uintptr_t MyDmacContext) {
-    static uint8_t servo = SERVO_1;
-
-    if (DMAC_TRANSFER_EVENT_COMPLETE == returned_evnt) {
-        bool overCurrent = false;
-        uint16_t input_voltage = 0;
-        for (size_t sample = 0; sample < TRANSFER_SIZE; sample++) {
-            input_voltage += adc_result_array[sample];
-
-            // 2.5 V == 0 A
-            /*input_voltage =*/
-            /*    (float)(adc_result_array[sample] * ADC_VREF / 4095U - 2.5)
-             * /*/
-            /*    0.4;*/
-
-            /*printf(*/
-            /*    "ADC Count = 0x%03x, ADC Input Current = %d.%03d A "*/
-            /*    "\n\r",*/
-            /*    adc_result_array[sample], (int)input_voltage,*/
-            /*    (int)((input_voltage - (int)input_voltage) * 100.0));*/
-        }
-        input_voltage = input_voltage / TRANSFER_SIZE;
-
-        if (input_voltage > VOLTAGE_THRESHOLD) {
-            overCurrent = true;
-        }
-        switch (servo) {
-            case SERVO_1:
-                if (overCurrent) {
-                    PORT_REGS->GROUP[0].PORT_OUTCLR |= (1 << 27);
-                }
-                ADC0_REGS->ADC_INPUTCTRL = ADC_POSINPUT_AIN1;
-                servo = SERVO_2;
-                break;
-            case SERVO_2:
-                if (overCurrent) {
-                    PORT_REGS->GROUP[0].PORT_OUTCLR |= (1 << 28);
-                }
-                ADC0_REGS->ADC_INPUTCTRL = ADC_POSINPUT_AIN4;
-                servo = SERVO_3;
-                break;
-            case SERVO_3:
-                if (overCurrent) {
-                    PORT_REGS->GROUP[0].PORT_OUTCLR |= (1 << 0);
-                }
-                ADC0_REGS->ADC_INPUTCTRL = ADC_POSINPUT_AIN0;
-                servo = SERVO_1;
-                break;
-            default:
-                break;
-        }
-
-        DMAC_ChannelTransfer(
-            DMAC_CHANNEL_0, (const void*)&ADC0_REGS->ADC_RESULT,
-            (const void*)adc_result_array, sizeof(adc_result_array));
-    } else if (DMAC_TRANSFER_EVENT_ERROR == returned_evnt) {
-        DMAC_ChannelTransfer(
-            DMAC_CHANNEL_0, (const void*)&ADC0_REGS->ADC_RESULT,
-            (const void*)adc_result_array, sizeof(adc_result_array));
-    }
-}
-
-
-static void callback_init(void){
-    CAN0_RxCallbackRegister(can_rx_callback, STATE_CAN_RECEIVE,
-                            CAN_MSG_ATTR_RX_FIFO0);
-    DMAC_ChannelCallbackRegister(DMAC_CHANNEL_0, dmac_channel0_callback, 0);
-    TC0_TimerCallbackRegister(tc0_callback, (uintptr_t)NULL);
-    TC1_TimerCallbackRegister(tc1_callback, (uintptr_t)NULL);
-}
